Moved compressed image tests from test_image.py to test_compressed.py
